+++ Contents/Code/PAsearchSites.py
@@ -82,7 +82,6 @@ import networkIntersec
 import networkCherryPimps
 import siteWicked
 import site18OnlyGirls
-import siteGirlsWay
 import siteEvilAngel
 import siteClubSeventeen
 import siteElegantAngel
@@ -96,7 +95,6 @@ import networkR18
 import networkBang
 import siteVivid
 import siteJaysPOV
-import siteGirlfriendsFilms
 import siteBAMVisions
 import sitePJGirls
 import networkPureCFNM
@@ -289,7 +287,7 @@ searchSites[179] = ["Tranny Surprise","Tranny Surprise", "https://www.realitykin
 searchSites[180] = ["VIP Crew","VIP Crew", "https://www.realitykings.com", "https://site-api.project1service.com"]
 searchSites[181] = ["We Live Together","We Live Together", "https://www.realitykings.com", "https://site-api.project1service.com"]
 searchSites[182] = ["Wives in Pantyhose","Wives in Pantyhose", "https://www.realitykings.com", "https://site-api.project1service.com"]
-searchSites[183] = ["21Naturals","21Naturals","https://www.21naturals.com","http://www.21naturals.com/en/search/"]
+searchSites[183] = ["21Naturals","21Naturals","https://www.21naturals.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[184] = ["PornFidelity","PornFidelity","https://www.pornfidelity.com","https://www.pornfidelity.com/episodes/search/?site=2&page=1&search="]
 searchSites[185] = ["TeenFidelity","TeenFidelity","https://www.pornfidelity.com","https://www.pornfidelity.com/episodes/search/?site=3&page=1&search="]
 searchSites[186] = ["Kelly Madison","Kelly Madison","https://www.pornfidelity.com","https://www.pornfidelity.com/episodes/search/?site=1&page=1&search="]
@@ -384,16 +382,16 @@ searchSites[274] = ["Elegant Anal","Elegant Anal","https://www.babes.com","https
 searchSites[275] = ["Office Obsession","Office Obsession","https://www.babes.com","https://site-api.project1service.com"]
 searchSites[276] = ["Stepmom Lessons","Stepmom Lessons","https://www.babes.com","https://site-api.project1service.com"]
 searchSites[277] = ["Evil Angel","Evil Angel","https://www.evilangel.com","https://www.evilangel.com/en/"]
-searchSites[278] = ["HardX","HardX","https://www.xempire.com","https://www.xempire.com/en/search/hardx/"]
+searchSites[278] = ["HardX","HardX","https://www.xempire.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[279] = ["GloryHoleSecrets","GloryHoleSecrets","http://www.gloryholesecrets.com","http://www.gloryholesecrets.com/tour/search.php?query="]
 searchSites[280] = ["New Sensations","New Sensations","http://www.newsensations.com","https://www.newsensations.com/tour_ns/"]
-searchSites[281] = ["Pure Taboo","Pure Taboo","https://www.puretaboo.com","https://www.puretaboo.com/en/search/"]
+searchSites[281] = ["Pure Taboo","Pure Taboo","https://www.puretaboo.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[282] = ["Swallowed","Swallowed","https://tour.swallowed.com","https://tour.swallowed.com/view/"]
 searchSites[283] = ["TrueAnal","TrueAnal","https://tour.trueanal.com","https://tour.trueanal.com/view/"]
 searchSites[284] = ["Nympho","Nympho","https://tour.nympho.com","https://tour.nympho.com/view/"]
-searchSites[285] = ["EroticaX","EroticaX","https://www.xempire.com","https://www.xempire.com/en/search/eroticax/"]
-searchSites[286] = ["DarkX","DarkX","https://www.xempire.com","https://www.xempire.com/en/search/darkx/"]
-searchSites[287] = ["LesbianX","LesbianX","http://www.xempire.com","http://www.xempire.com/en/search/lesbianx/"]
+searchSites[285] = ["EroticaX","EroticaX","https://www.xempire.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[286] = ["DarkX","DarkX","https://www.xempire.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[287] = ["LesbianX","LesbianX","http://www.xempire.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[288] = ["Twistys","Twistys","https://www.twistys.com","https://site-api.project1service.com"]
 searchSites[289] = ["WhenGirlsPlay","WhenGirlsPlay","https://www.twistys.com","https://site-api.project1service.com"]
 searchSites[290] = ["MomKnowsBest","MomKnowsBest","https://www.twistys.com","https://site-api.project1service.com"]
@@ -435,9 +433,9 @@ searchSites[325] = ["CockCompetition","CockCompetition","https://www.pornpros.co
 searchSites[326] = ["PimpParade","PimpParade","https://www.pornpros.com","https://pornpros.com/video/"]
 searchSites[327] = ["SquirtDisgrace","SquirtDisgrace","https://www.pornpros.com","https://pornpros.com/video/"]
 searchSites[328] = ["DigitalPlayground","DigitalPlayground","https://www.digitalplayground.com","https://site-api.project1service.com"]
-searchSites[329] = ["Throated","Throated","https://www.blowpass.com","https://www.blowpass.com/en/search/throated/scene/"]
-searchSites[330] = ["Nuru Massage","Nuru Massage","https://www.fantasymassage.com","https://www.fantasymassage.com/en/search/"]
-searchSites[331] = ["Nuru Massage","Nuru Massage","https://www.fantasymassage.com","https://www.fantasymassage.com/en/search/"]
+searchSites[329] = ["Throated","Throated","https://www.blowpass.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[330] = ["Nuru Massage","Nuru Massage","https://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+
 searchSites[332] = ["DDF Babes", "DDF Babes", "https://ddfnetwork.com","https://ddfnetwork.com/videos/freeword/"]
 searchSites[333] = ["SexyHub","SexyHub","https://www.sexyhub.com","https://site-api.project1service.com"]
 searchSites[334] = ["Dane Jones","Dane Jones","https://www.sexyhub.com","https://site-api.project1service.com"]
@@ -457,48 +455,48 @@ searchSites[347] = ["Bad Daddy POV","Bad Daddy POV","https://www.baddaddypov.com
 searchSites[348] = ["POV Perverts","POV Perverts","https://www.povperverts.net","https://povperverts.net/search.php?query="]
 searchSites[349] = ["Pervert Gallery","Pervert Gallery","https://www.pervertgallery.com","https://pervertgallery.com/search.php?query="]
 searchSites[350] = ["DTF Sluts","DTF Sluts","https://www.dtfsluts.com","https://dtfsluts.com/search.php?query="]
-searchSites[351] = ["Mommy Blows Best","Mommy Blows Best","http://www.blowpass.com","http://www.blowpass.com/en/search/mommyblowsbest/scene/"]
-searchSites[352] = ["Only Teen Blowjobs","Only Teen Blowjobs","http://www.blowpass.com","http://www.blowpass.com/en/search/onlyteenblowjobs/scene/"]
-searchSites[353] = ["1000 Facials","1000 Facials","http://www.blowpass.com","http://www.blowpass.com/en/search/1000facials/scene/"]
-searchSites[354] = ["Immoral Live","Immoral Live","http://www.blowpass.com","http://www.blowpass.com/en/search/immorallive/scene/"]
-searchSites[355] = ["Fantasy Massage","Fantasy Massage","http://www.fantasymassage.com","http://www.fantasymassage.com/en/search/"]
-searchSites[356] = ["All Girl Massage","All Girl Massage","http://www.fantasymassage.com","http://www.fantasymassage.com/en/search/"]
-searchSites[357] = ["Soapy Massage","Soapy Massage","http://www.fantasymassage.com","http://www.fantasymassage.com/en/search/"]
-searchSites[358] = ["Milking Table","Milking Table","http://www.fantasymassage.com","http://www.fantasymassage.com/en/search/"]
-searchSites[359] = ["Massage Parlor","Massage Parlor","http://www.fantasymassage.com","http://www.fantasymassage.com/en/search/"]
-searchSites[360] = ["Tricky Spa","Tricky Spa","http://www.fantasymassage.com","http://www.fantasymassage.com/en/search/"]
+searchSites[351] = ["Mommy Blows Best","Mommy Blows Best","http://www.blowpass.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[352] = ["Only Teen Blowjobs","Only Teen Blowjobs","http://www.blowpass.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[353] = ["1000 Facials","1000 Facials","http://www.blowpass.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[354] = ["Immoral Live","Immoral Live","http://www.blowpass.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[355] = ["Fantasy Massage","Fantasy Massage","http://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[356] = ["All Girl Massage","All Girl Massage","http://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[357] = ["Soapy Massage","Soapy Massage","http://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[358] = ["Milking Table","Milking Table","http://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[359] = ["Massage Parlor","Massage Parlor","http://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[360] = ["Tricky Spa","Tricky Spa","http://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[361] = ["Sweetheart Video","Sweetheart Video","https://www.milehighmedia.com","https://site-api.project1service.com"]
 searchSites[362] = ["Reality Junkies","Reality Junkies","http://www.milehighmedia.com","https://site-api.project1service.com"]
 searchSites[363] = ["SweetSinner","SweetSinner","https://www.milehighmedia.com","https://site-api.project1service.com"]
 searchSites[364] = ["Doghouse Digital","Doghouse Digital","http://www.milehighmedia.com","https://site-api.project1service.com"]
-searchSites[365] = ["21Sextury","21Sextury","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[366] = ["Anal Teen Angels","Anal Teen Angels","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[367] = ["Deepthroat Frenzy","Deepthroat Frenzy","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[368] = ["DP Fanatics","DP Fanatics","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[369] = ["Footsie Babes","Footsie Babes","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[370] = ["Gapeland","Gapeland","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[371] = ["Lez Cuties","Lez Cuties","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[372] = ["Pix and Video","Pix and Video","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
-searchSites[373] = ["21FootArt","21FootArt","http://www.21naturals.com","http://www.21naturals.com/en/search/"]
-searchSites[374] = ["21EroticAnal","21EroticAnal","http://www.21naturals.com","http://www.21naturals.com/en/search/"]
+searchSites[365] = ["21Sextury","21Sextury","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[366] = ["Anal Teen Angels","Anal Teen Angels","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[367] = ["Deepthroat Frenzy","Deepthroat Frenzy","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[368] = ["DP Fanatics","DP Fanatics","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[369] = ["Footsie Babes","Footsie Babes","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[370] = ["Gapeland","Gapeland","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[371] = ["Lez Cuties","Lez Cuties","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[372] = ["Pix and Video","Pix and Video","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[373] = ["21FootArt","21FootArt","http://www.21naturals.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[374] = ["21EroticAnal","21EroticAnal","http://www.21naturals.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[375] = ["Mommys Girl","Mommy's Girl","http://www.girlsway.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[376] = ["Web Young","Web Young","http://www.girlsway.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[377] = ["Girls Try Anal","Girls Try Anal","http://www.girlsway.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[378] = ["Sextape Lesbians","Sextape Lesbians","http://www.girlsway.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[379] = ["Girlsway Originals","Girlsway Originals","http://www.girlsway.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
-searchSites[380] = ["Girlfriends Films","Girlfriends Films","http://www.girlfriendsfilms.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/query"]
-searchSites[381] = ["Burning Angel","Burning Angel","http://www.burningangel.com","http://www.burningangel.com/en/search/"]
-searchSites[382] = ["Pretty Dirty","Pretty Dirty","http://www.prettydirty.com","http://www.prettydirty.com/en/search/"]
-searchSites[383] = ["Devils Film","Devil's Film","http://www.devilsfilm.com","http://www.devilsfilm.com/en/search/"]
-searchSites[384] = ["Peter North","Peter North","http://www.peternorth.com","http://www.peternorth.com/en/search/"]
-searchSites[385] = ["Rocco Siffredi","Rocco Siffredi","http://www.roccosiffredi.com","http://www.roccosiffredi.com/en/search/"]
-searchSites[386] = ["Tera Patrick","Tera Patrick","http://www.terapatrick.com","http://www.terapatrick.com/en/search/"]
-searchSites[387] = ["Sunny Leone","Sunny Leone","http://www.sunnyleone.com","http://www.sunnyleone.com/en/search/scene/"]
-searchSites[388] = ["Lane Sisters","Lane Sisters","http://www.lanesisters.com","http://www.lanesisters.com/en/search/scene/"]
-searchSites[389] = ["Dylan Ryder","Dylan Ryder","http://www.dylanryder.com","http://www.dylanryder.com/en/search/scene/"]
-searchSites[390] = ["Abbey Brooks","Abbey Brooks","http://www.abbeybrooks.com","http://www.abbeybrooks.com/en/search/scene/"]
-searchSites[391] = ["Devon Lee","Devon Lee","http://www.devonlee.com","http://www.devonlee.com/en/search/scene/"]
-searchSites[392] = ["Hanna Hilton","Hanna Hilton","http://www.hannahilton.com","http://www.hannahilton.com/en/search/scene/"]
+searchSites[380] = ["Girlfriends Films","Girlfriends Films","http://www.girlfriendsfilms.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[381] = ["Burning Angel","Burning Angel","http://www.burningangel.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[382] = ["Pretty Dirty","Pretty Dirty","http://www.prettydirty.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[383] = ["Devils Film","Devil's Film","http://www.devilsfilm.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[384] = ["Peter North","Peter North","http://www.peternorth.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[385] = ["Rocco Siffredi","Rocco Siffredi","http://www.roccosiffredi.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[386] = ["Tera Patrick","Tera Patrick","http://www.terapatrick.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[387] = ["Sunny Leone","Sunny Leone","http://www.sunnyleone.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[388] = ["Lane Sisters","Lane Sisters","http://www.lanesisters.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[389] = ["Dylan Ryder","Dylan Ryder","http://www.dylanryder.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[390] = ["Abbey Brooks","Abbey Brooks","http://www.abbeybrooks.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[391] = ["Devon Lee","Devon Lee","http://www.devonlee.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[392] = ["Hanna Hilton","Hanna Hilton","http://www.hannahilton.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[393] = ["LA Sluts","LA Sluts","https://tour.naughtyamerica.com","https://tour.naughtyamerica.com/search?term="]
 searchSites[394] = ["Slut Stepsister","Slut Stepsister","https://tour.naughtyamerica.com","https://tour.naughtyamerica.com/search?term="]
 searchSites[395] = ["Teens Love Cream","Teens Love Cream","https://tour.naughtyamerica.com","https://tour.naughtyamerica.com/search?term="]
@@ -566,13 +564,13 @@ searchSites[456] = ["GiveMePink","GiveMePink", "https://www.perfectgonzo.com","h
 searchSites[457] = ["FistFlush","FistFlush", "https://www.perfectgonzo.com","https://www.perfectgonzo.com/movies?tag=fistflush&q="]
 searchSites[458] = ["MilfThing","MilfThing", "https://www.perfectgonzo.com","https://www.perfectgonzo.com/movies?tag=milfthing&q="]
 searchSites[459] = ["PerfectGonzoInterview","PerfectGonzoInterview", "https://www.perfectgonzo.com","https://www.perfectgonzo.com/movies?tag=interview&q="]
-searchSites[460] = ["21Sextreme","21Sextreme","http://www.21sextreme.com","http://www.21sextreme.com/en/search/"]
-searchSites[461] = ["LustyGrandmas","LustyGrandmas","http://www.21sextreme.com","http://www.21sextreme.com/en/search/"]
-searchSites[462] = ["GrandpasFuckTeens","GrandpasFuckTeens","http://www.21sextreme.com","http://www.21sextreme.com/en/search/"]
-searchSites[463] = ["TeachMeFisting","TeachMeFisting","http://www.21sextreme.com","http://www.21sextreme.com/en/search/"]
-searchSites[464] = ["Zoliboy","Zoliboy","http://www.21sextreme.com","http://www.21sextreme.com/en/search/"]
-searchSites[465] = ["DominatedGirls","DominatedGirls","http://www.21sextreme.com","http://www.21sextreme.com/en/search/"]
-searchSites[466] = ["Asshole Fever","Asshole Fever","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
+searchSites[460] = ["21Sextreme","21Sextreme","http://www.21sextreme.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[461] = ["LustyGrandmas","LustyGrandmas","http://www.21sextreme.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[462] = ["GrandpasFuckTeens","GrandpasFuckTeens","http://www.21sextreme.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[463] = ["TeachMeFisting","TeachMeFisting","http://www.21sextreme.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[464] = ["Zoliboy","Zoliboy","http://www.21sextreme.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[465] = ["DominatedGirls","DominatedGirls","http://www.21sextreme.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
+searchSites[466] = ["Asshole Fever","Asshole Fever","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[467] = ["Anal College","Anal College","https://tour.naughtyamerica.com","https://tour.naughtyamerica.com/search?term="]
 searchSites[468] = ["Watch Your Wife","Watch Your Wife","https://tour.naughtyamerica.com","https://tour.naughtyamerica.com/search?term="]
 searchSites[469] = ["BadoinkVR","BadoinkVR","https://www.badoinkvr.com","https://badoinkvr.com/vrpornvideos/search/"]
@@ -798,7 +796,7 @@ searchSites[688] = ["Hegre","Hegre","http://www.hegre.com","http://www.hegre.com
 searchSites[689] = ["Femdom Empire","Femdom Empire","https://femdomempire.com","https://femdomempire.com/tour/search.php?st=advanced&qany="]
 searchSites[690] = ["Day With A Porn Star","Day With A Porn Star","http://www.brazzers.com","http://www.brazzers.com/search/all/?q="]
 searchSites[691] = ["Watch Your Mom","Watch Your Mom","https://tour.naughtyamerica.com","https://tour.naughtyamerica.com/search?term="]
-searchSites[692] = ["Butt Plays","Butt Plays","http://www.21sextury.com","http://www.21sextury.com/en/search/"]
+searchSites[692] = ["Butt Plays","Butt Plays","http://www.21sextury.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[693] = ["Dorcel Vision","Dorcel Vision","https://www.dorcelvision.com","https://www.dorcelvision.com/en/search?type=4&keyword="]
 searchSites[694] = ["Feminized","Feminized","http://feminized.com","http://feminized.com/tour/search.php?st=advanced&qany="]
 searchSites[695] = ["XConfessions","XConfessions","https://xconfessions.com","https://2rzi1cnto2-dsn.algolia.net/1/indexes/*/queries"]
@@ -856,7 +854,7 @@ searchSites[746] = ["Girls Rimming","Girls Rimming","https://www.girlsrimming.co
 searchSites[747] = ["Gangbang Creampie","Gangbang Creampie","https://gangbangcreampie.com","https://gangbangcreampie.com/tour/search.php?query="]
 searchSites[748] = ["DadCrush","DadCrush","https://www.dadcrush.com","https://www.dadcrush.com/movies/"]
 searchSites[749] = ["Show My BF","Show My BF","https://tour.naughtyamerica.com","https://tour.naughtyamerica.com/search?term="]
-searchSites[750] = ["POV Massage","POV Massage","http://www.fantasymassage.com","http://www.fantasymassage.com/en/search/"]
+searchSites[750] = ["POV Massage","POV Massage","http://www.fantasymassage.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 searchSites[751] = ["Step Secrets","Step Secrets","http://www.stepsecrets.com","https://stepsecrets.com/?query="]
 searchSites[752] = ["VRHush","VRHush","https://www.vrhush.com","https://vrhush.com/scenes/"]
 searchSites[753] = ["MetArt","MetArt","https://www.metart.com","https://www.metart.com/api"]
@@ -949,7 +947,7 @@ searchSites[839] = ["Upskirt Jerk","Upskirt Jerk","https://www.upskirtjerk.com",
 searchSites[840] = ["Interracial Pass","Interracial Pass","hhttps://www.interracialpass.com","https://www.interracialpass.com/newtour/search.php?query="]
 searchSites[841] = ["LookAtHerNow","LookAtHerNow","https://www.lookathernow.com","https://site-api.project1service.com"]
 searchSites[842] = ["Mylfwood","Mylfwood","https://www.mylf.com","https://www.mylf.com/movies/"]
-searchSites[843] = ["AllBlackX","AllBlackX","https://www.xempire.com","https://www.xempire.com/en/search/allblackx/"]
+searchSites[843] = ["AllBlackX","AllBlackX","https://www.xempire.com","https://tsmkfa364q-dsn.algolia.net/1/indexes/*/queries"]
 
 def getSearchBaseURL(siteID):
     return searchSites[siteID][2]
